+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
-- 修复 `f-mack` 组件 `z-index` 属性丢失问题
-- `f-mask` 组件新增 `mask-background` 和 `mask-opacity` 配置项
+**重点更新**
+
+- `f-mask` 组件现已更名为 `f-popup`
+
+**修复 Bug**
+
+- 修复 `f-popup` 组件 `z-index` 属性丢失问题
+- `f-popup` 组件新增 `mask-background` 和 `mask-opacity` 配置项
 
 ## 0.7.0-alpha.5 (2022-10-13)